자잘한 버그 수정
\ㅇ-ㅇ\
