--- WhereIsMyHome_Final/src/main/java/com/mycom/myhome/address/PropertyDao.java
@@ -1,13 +1,0 @@
-package com.mycom.myhome.address;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface PropertyDao {
-	public List<PropertyDto> searchByDong(String dongName);
-	public List<PropertyDto> searchByGugunCode(String gugunCode);
-	public List<PropertyDto> searchBySidoCode(String sidoCode);
-	public List<PropertyDto> searchByApt(String aptName);
-}
