--- WhereIsMyHome_Final/src/main/java/com/mycom/myhome/address/PropertyDetailDao.java
@@ -1,10 +1,0 @@
-package com.mycom.myhome.address;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface PropertyDetailDao {
-	public List<PropertyDetailDto> propertyDetail(String houseNo);
-}
