--- WhereIsMyHome_Final/src/main/java/com/mycom/myhome/address/PropertyDetailDto.java
@@ -1,24 +1,0 @@
-package com.mycom.myhome.address;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class PropertyDetailDto {
-	private int no;
-	private String dong;
-	private String aptName;
-	private String code;
-	private String dealAmount;
-	private String buildYear;
-	private String dealYear;
-	private String dealMonth;
-	private String dealDay;
-	private String area;
-	private String floor;
-	private String jibun;
-	private String houseNo;
-}
