--- WhereIsMyHome_Final/src/main/java/com/mycom/myhome/address/PropertyDto.java
@@ -1,19 +1,0 @@
-package com.mycom.myhome.address;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class PropertyDto {
-	private String no;
-	private String dong;
-	private String aptName;
-	private String code;
-	private String buildYear;
-	private String jibun;
-	private String lat;
-	private String lng;
-}
