--- WhereIsMyHome_Final/src/main/java/com/mycom/myhome/address/StationController.java
@@ -1,26 +1,0 @@
-package com.mycom.myhome.address;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@CrossOrigin(
-		origins = {"http://localhost:5500", "http://localhost:5501"},
-		allowCredentials = "true", 
-		allowedHeaders = "*", 
-		methods = {RequestMethod.GET,RequestMethod.POST,RequestMethod.DELETE,RequestMethod.PUT,RequestMethod.HEAD,RequestMethod.OPTIONS}
-	)
-public class StationController {
-	@Autowired
-	StationService service;
-	
-	@GetMapping("/stations")
-	private List<StationDto> stationList () {
-		return service.stationList();
-	}
-}
