--- WhereIsMyHome_Final/src/main/java/com/mycom/myhome/address/StationDto.java
@@ -1,16 +1,0 @@
-package com.mycom.myhome.address;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class StationDto {
-	String line;
-	String name;
-	String code;
-	String lat;
-	String lng;
-}
