--- WhereIsMyHome_Final/src/main/java/com/mycom/myhome/address/StationService.java
@@ -1,7 +1,0 @@
-package com.mycom.myhome.address;
-
-import java.util.List;
-
-public interface StationService {
-	public List<StationDto> stationList ();
-}
