--- WhereIsMyHome_Final/src/main/java/com/mycom/myhome/address/StationServiceImpl.java
@@ -1,23 +1,0 @@
-package com.mycom.myhome.address;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class StationServiceImpl implements StationService{
-	@Autowired
-	StationDao dao;
-
-	@Override
-	public List<StationDto> stationList() {
-		List<StationDto> list=null;
-		try {
-			list=dao.stationList();
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		return list;
-	}	
-}
