+++ WhereIsMyHome_Final/src/main/java/com/mycom/myhome/board/file/BoardFileResultDto.java
@@ -1,0 +1,24 @@
+package com.mycom.myhome.board.file;
+
+import com.mycom.myhome.common.Status;
+
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+@Data
+@RequiredArgsConstructor(staticName="of")
+public class BoardFileResultDto {
+
+	private final Status result;
+	private final String message;
+	
+	private final BoardFile file;
+	
+	public static BoardFileResultDto ofSuccess(String message, BoardFile file) {
+		return new BoardFileResultDto(Status.SUCCESS, message, file);
+	}
+	
+	public static BoardFileResultDto ofFail(String message) {
+		return new BoardFileResultDto(Status.FAIL, message, null);
+	}
+}
