+++ WhereIsMyHome_Final/src/main/java/com/mycom/myhome/property/PropertyDetailDao.java
@@ -1,0 +1,10 @@
+package com.mycom.myhome.property;
+
+import java.util.List;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface PropertyDetailDao {
+	public List<PropertyDetailDto> propertyDetail(String houseNo);
+}
