+++ WhereIsMyHome_Final/src/main/java/com/mycom/myhome/property/PropertyDto.java
@@ -1,0 +1,19 @@
+package com.mycom.myhome.property;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class PropertyDto {
+	private String no;
+	private String dong;
+	private String aptName;
+	private String code;
+	private String buildYear;
+	private String jibun;
+	private String lat;
+	private String lng;
+}
