+++ WhereIsMyHome_Final/src/main/java/com/mycom/myhome/station/StationDao.java
@@ -1,0 +1,10 @@
+package com.mycom.myhome.station;
+
+import java.util.List;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface StationDao {
+	public List<StationDto> stationList ();
+}
