+++ WhereIsMyHome_Final/src/main/java/com/mycom/myhome/station/StationService.java
@@ -1,0 +1,7 @@
+package com.mycom.myhome.station;
+
+import java.util.List;
+
+public interface StationService {
+	public List<StationDto> stationList ();
+}
