+++ WhereIsMyHome_Final/src/main/resources/mapper/board_query.xml
@@ -22,7 +22,7 @@
 		FROM board 
 		WHERE board_id = #{boardId}
 	</select>
-	<select id="getFileList" parameterType="int" resultType="com.mycom.myhome.board.BoardFile">
+	<select id="getFileList" parameterType="int" resultType="com.mycom.myhome.board.file.BoardFile">
 		SELECT file_id, board_id, file_name, file_size, file_content_type, file_url, reg_dt
 		FROM board_file
 		WHERE board_id = #{boardId}
@@ -45,11 +45,11 @@
 		INSERT INTO board(user_seq, writer, title, content, reg_dt, read_count, board_clsf)
 		VALUES(#{userSeq}, #{writer}, #{title}, #{content}, now(), #{readCount}, #{boardClsf})
 	</insert>
-	<insert id="uploadFile" parameterType="com.mycom.myhome.board.BoardFile">
+	<insert id="uploadFile" parameterType="com.mycom.myhome.board.file.BoardFile">
 		INSERT INTO board_file(board_id, file_name, file_size, file_content_type, file_url, reg_dt)
 		VALUES(#{boardId}, #{fileName}, #{fileSize}, #{fileContentType}, #{fileUrl}, now())
 	</insert>
-	<select id="downloadFile" parameterType="int" resultType="com.mycom.myhome.board.BoardFile">
+	<select id="downloadFile" parameterType="int" resultType="com.mycom.myhome.board.file.BoardFile">
 		SELECT file_id, board_id, file_name, file_size, file_content_type, file_url, reg_dt
 		FROM board_file
 		WHERE file_id = #{fileId}
@@ -87,7 +87,7 @@
 		FROM interest_area
 		WHERE user_seq = #{userSeq}
 	</select>
-	<select id="findByHouseNo" parameterType="int" resultType="com.mycom.myhome.address.PropertyDetailDto">
+	<select id="findByHouseNo" parameterType="int" resultType="com.mycom.myhome.property.PropertyDetailDto">
 		SELECT dong, aptname, buildyear, code,(
 			SELECT dealAmount
 			FROM housedeal
@@ -96,6 +96,15 @@
 		FROM houseinfo hi
 		WHERE no = #{houseNo}
 	</select>
+	<insert id="insertInterestArea" parameterType="map">
+		INSERT INTO interest_area(house_no, user_seq)
+		VALUES(#{houseNo}, #{userSeq})
+	</insert>
+	<delete id="deleteByNoSeq" parameterType="map">
+		DELETE FROM interest_area
+		WHERE house_no = #{houseNo} 
+			and user_seq = #{userSeq}
+	</delete>
 </mapper>
