+++ WhereIsMyHome_Final/src/main/resources/mapper/station_query.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.mycom.myhome.address.StationDao"> <!--  @Mapper annotation 있는 dao -->
+<mapper namespace="com.mycom.myhome.station.StationDao"> <!--  @Mapper annotation 있는 dao -->
 	<!-- 동 이름으로 집 정보 받아오기-->
-	<select id="stationList" resultType="com.mycom.myhome.address.StationDto">
+	<select id="stationList" resultType="com.mycom.myhome.station.StationDto">
 		select line, name, code, lat, lng from station
 	</select>
 </mapper>
